% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_deeploc.R
\name{generate_deeploc_fasta}
\alias{generate_deeploc_fasta}
\title{Generate FASTA for DeepLoc}
\usage{
generate_deeploc_fasta(sce, species = "Human", output_file = "proteins.fasta")
}
\arguments{
\item{sce}{SingleCellExperiment object.}

\item{species}{"Human" or "Mouse".}

\item{output_file}{Path to save the FASTA file.}
}
\description{
Creates a FASTA file of protein sequences from the SCE object.
}
